register Binding as soon as it's declared
With the old approach - keeping a list of classes and registering them in the ``ready``-method in ``apps.py`` - I had the problem that ready got called before my ``consumers.py``, where my bindings are, was loaded. So ``register_all`` ran on an empty list and my bindings where not registered.  I also added a uid to the signal-connections to prevent duplicate messages.