Fixed bug with _binding_group_names
Since the old_group_names were saved as an attribute of ``instance`` in the ``pre_change_receiver`` it did not carry over to ``post_change_receiver`` and an ``AttributeError`` was raised. I think that saving ``old_group_names`` as a class attribute of the binding is the best fix for this.