Remove 'onopen' for compatibility and usability
socket.onopen doesn't work as expected for me on chromium based browsers (chrome and vivaldi). Instead, I recommend suggesting a straight send command in the tutorial. In addition to being more compatible, this also is more directly actionable (user can try .send over and over again, opening browsers of different types, etc.) A small change, but this is what I ended up doing, and I was stuck for a few with the .onopen attempts.