Relax consumer_callable requirements
According to ASGI specification, both "receive" and "send" should be
an awaitable callable, where as create_task(), accept only coroutine.

NGINX Unit ASGI server implementation, provides "receive" and "send"
callable which returns "Future", which in turns is awaitable and can
be used just as a "Task".

Using ensure_future() instead of create_task() is a simple way to create
"Task" for coroutines and use "Future" if it is already provided.