postpone model import to avoid ImproperlyConfigured error
import of channels.auth is triggering an import of model, which
should not happen before ASGI app is ready.
Doing so raises an exception which might be tricky for newcomers
to solve. To make it work, code needs to be mixed among imports.
Postponing model import solves the common culprit of this error.